Trait rustc_data_structures::aligned::Aligned

pub unsafe trait Aligned {
    const ALIGN: Alignment;
}

A type with a statically known alignment.

Safety

Self::ALIGN must be equal to the alignment of Self. For sized types it is mem::align_of<Self>(), for unsized types it depends on the type, for example [T] has alignment of T.

Required Associated Constants

const ALIGN: Alignment

Alignment of Self.