Struct rustc_data_structures::bit_set::BitSet

pub struct BitSet<T: Idx> { /* fields omitted */ }

A fixed-size bitset type with a dense representation. It does not support resizing after creation; use GrowableBitSet for that.

T is an index type, typically a newtyped usize wrapper, but it can also just be usize.

All operations that involve an element will panic if the element is equal to or greater than the domain size. All operations that involve two bitsets will panic if the bitsets have differing domain sizes.

Methods

impl<T: Idx> BitSet<T>

pub fn new_empty(domain_size: usize) -> BitSet<T>

Creates a new, empty bitset with a given domain_size.

pub fn new_filled(domain_size: usize) -> BitSet<T>

Creates a new, filled bitset with a given domain_size.

pub fn domain_size(&self) -> usize

Gets the domain size.

pub fn clear(&mut self)

Clear all elements.

pub fn overwrite(&mut self, other: &BitSet<T>)

Efficiently overwrite self with other.

pub fn count(&self) -> usize

Count the number of set bits in the set.

pub fn contains(&self, elem: T) -> bool

Returns true if self contains elem.

pub fn superset(&self, other: &BitSet<T>) -> bool

Is self is a (non-strict) superset of other?

pub fn is_empty(&self) -> bool

Is the set empty?

pub fn insert(&mut self, elem: T) -> bool

Insert elem. Returns whether the set has changed.

pub fn insert_all(&mut self)

Sets all bits to true.

pub fn remove(&mut self, elem: T) -> bool

Returns true if the set has changed.

pub fn union(&mut self, other: &impl UnionIntoBitSet<T>) -> bool

Sets self = self | other and returns true if self changed (i.e., if new bits were added).

pub fn subtract(&mut self, other: &impl SubtractFromBitSet<T>) -> bool

Sets self = self - other and returns true if self changed. (i.e., if any bits were removed).

pub fn intersect(&mut self, other: &BitSet<T>) -> bool

Sets self = self & other and return true if self changed. (i.e., if any bits were removed).

Important traits for &'_ [u8]
impl<'_> Read for &'_ [u8]impl<'_> Write for &'_ mut [u8]

pub fn words(&self) -> &[Word]

Gets a slice of the underlying words.

Important traits for BitIter<'a, T>
impl<'a, T: Idx> Iterator for BitIter<'a, T> type Item = T;

pub fn iter<'a>(&'a self) -> BitIter<'a, T>

Iterates over the indices of set bits in a sorted order.

pub fn to_hybrid(&self) -> HybridBitSet<T>

Duplicates the set as a hybrid set.