Struct rustc_data_structures::bitslice::BitLookup

struct BitLookup {
    word: usize,
    bit_in_word: usize,
    bit_mask: Word,
}

Fields
word: usize

An index of the word holding the bit in original [Word] of query.

bit_in_word: usize

Index of the particular bit within the word holding the bit.

bit_mask: Word

Word with single 1-bit set corresponding to where the bit is located.