Module rustc_data_structures::bitvec

Structs

BitMatrix
A "bit matrix" is basically a matrix of booleans represented as one gigantic bitvector. In other words, it is as if you have rows bitvectors, each of length columns.

BitVector
A very simple BitVector type.

BitVectorIter

Functions

u64s

word_mask