Struct rustc_data_structures::control_flow_graph::dominators::DominatorTree

pub struct DominatorTree<N: Idx> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<Node: Idx> DominatorTree<Node>

pub fn children(&self, node: Node) -> &[Node]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?