Struct rustc_data_structures::flock::Lock

pub struct Lock { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl Lock

pub fn new(p: &Path, wait: bool, create: bool, exclusive: bool) -> Result<Lock>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

impl Lock

pub fn panicking_new(
p: &Path,
wait: bool,
create: bool,
exclusive: bool
) -> Lock

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?