rustc_data_structures

Module fx

Structs

FxHasher

A speedy hash algorithm for use within rustc. The hashmap in liballoc by default uses SipHash which isn’t quite as speedy as we want. In the compiler we’re not really worried about DOS attempts, so we use a fast non-cryptographic hash.

Traits

MutableValues

Opt-in mutable access to IndexSet values.

Type Aliases

FxHashMap

Type alias for a hash map that uses the Fx hashing algorithm.

FxHashSet

Type alias for a hash set that uses the Fx hashing algorithm.

FxIndexMap

FxIndexSet

IndexEntry

IndexOccupiedEntry

StdEntry