Module rustc_data_structures::graph

Modules

dominators

Finding the dominators in a control-flow graph.

implementation

A graph module for use in dataflow, region resolution, and elsewhere.

iterate

reference 🔒

scc

Routine to compute the strongly connected components (SCCs) of a graph.

vec_graph

Traits

ControlFlowGraph

DirectedGraph

GraphPredecessors

GraphSuccessors

WithNumEdges

WithNumNodes

WithPredecessors

WithStartNode

WithSuccessors

Functions

is_cyclic

Returns true if the graph has a cycle that is reachable from the start node.