Struct rustc_data_structures::graph::Graph

pub struct Graph<N, E> { /* fields omitted */ }

Unstable (rustc_private)

Methods

impl<N: Debug, E: Debug> Graph<N, E>

fn new() -> Graph<N, E>

Unstable (rustc_private)

fn all_nodes(&self) -> &[Node<N>]

Unstable (rustc_private)

fn len_nodes(&self) -> usize

Unstable (rustc_private)

fn all_edges(&self) -> &[Edge<E>]

Unstable (rustc_private)

fn len_edges(&self) -> usize

Unstable (rustc_private)

fn next_node_index(&self) -> NodeIndex

Unstable (rustc_private)

fn add_node(&mut self, data: N) -> NodeIndex

Unstable (rustc_private)

fn mut_node_data(&mut self, idx: NodeIndex) -> &mut N

Unstable (rustc_private)

fn node_data(&self, idx: NodeIndex) -> &N

Unstable (rustc_private)

fn node(&self, idx: NodeIndex) -> &Node<N>

Unstable (rustc_private)

fn next_edge_index(&self) -> EdgeIndex

Unstable (rustc_private)

fn add_edge(&mut self, source: NodeIndex, target: NodeIndex, data: E) -> EdgeIndex

Unstable (rustc_private)

fn mut_edge_data(&mut self, idx: EdgeIndex) -> &mut E

Unstable (rustc_private)

fn edge_data(&self, idx: EdgeIndex) -> &E

Unstable (rustc_private)

fn edge(&self, idx: EdgeIndex) -> &Edge<E>

Unstable (rustc_private)

fn first_adjacent(&self, node: NodeIndex, dir: Direction) -> EdgeIndex

Unstable (rustc_private)

Accesses the index of the first edge adjacent to node. This is useful if you wish to modify the graph while walking the linked list of edges.

fn next_adjacent(&self, edge: EdgeIndex, dir: Direction) -> EdgeIndex

Unstable (rustc_private)

Accesses the next edge in a given direction. This is useful if you wish to modify the graph while walking the linked list of edges.

fn each_node<'a, F>(&'a self, f: F) -> bool where F: FnMut(NodeIndex, &'a Node<N>) -> bool

Unstable (rustc_private)

Iterates over all edges defined in the graph.

fn each_edge<'a, F>(&'a self, f: F) -> bool where F: FnMut(EdgeIndex, &'a Edge<E>) -> bool

Unstable (rustc_private)

Iterates over all edges defined in the graph

fn outgoing_edges(&self, source: NodeIndex) -> AdjacentEdges<N, E>

Unstable (rustc_private)

fn incoming_edges(&self, source: NodeIndex) -> AdjacentEdges<N, E>

Unstable (rustc_private)

fn adjacent_edges(&self, source: NodeIndex, direction: Direction) -> AdjacentEdges<N, E>

Unstable (rustc_private)

fn successor_nodes(&self, source: NodeIndex) -> AdjacentTargets<N, E>

Unstable (rustc_private)

fn predecessor_nodes(&self, target: NodeIndex) -> AdjacentSources<N, E>

Unstable (rustc_private)

fn iterate_until_fixed_point<'a, F>(&'a self, op: F) where F: FnMut(usize, EdgeIndex, &'a Edge<E>) -> bool

Unstable (rustc_private)

fn depth_traverse<'a>(&'a self, start: NodeIndex, direction: Direction) -> DepthFirstTraversal<'a, N, E>

Unstable (rustc_private)