Struct rustc_data_structures::graph::vec_graph::VecGraph

pub struct VecGraph<N: Idx> {
    node_starts: IndexVec<N, usize>,
    edge_targets: Vec<N>,
}

Fields
node_starts: IndexVec<N, usize>

Maps from a given node to an index where the set of successors for that node starts. The index indexes into the edges vector. To find the range for a given node, we look up the start for that node and then the start for the next node (i.e., with an index 1 higher) and get the range between the two. This vector always has an extra entry so that this works even for the max element.

edge_targets: Vec<N>
Implementations

impl<N: Idx + Ord> VecGraph<N>

pub fn new(num_nodes: usize, edge_pairs: Vec<(N, N)>) -> Self

pub fn successors(&self, source: N) -> &[N]

Gets the successors for source as a slice.