Struct rustc_data_structures::indexed_set::IdxSet

pub struct IdxSet<T: Idx> {
    _pd: PhantomData<fn(_: &T)>,
    bits: [Word],
}

Represents a set (or packed family of sets), of some element type E, where each E is identified by some unique index type T.

In other words, T is the type used to index into the bitslice this type uses to represent the set of object it holds.

Fields
_pd: PhantomData<fn(_: &T)> bits: [Word]
Methods

impl<T: Idx> IdxSet<T>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

unsafe fn from_slice(s: &[Word]) -> &Self

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

unsafe fn from_slice_mut(s: &mut [Word]) -> &mut Self

impl<T: Idx> IdxSet<T>

pub fn to_owned(&self) -> IdxSetBuf<T>

pub fn clear(&mut self)

Removes all elements

pub fn remove(&mut self, elem: &T) -> bool

Removes elem from the set self; returns true iff this changed self.

pub fn add(&mut self, elem: &T) -> bool

Adds elem to the set self; returns true iff this changed self.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn range(&self, elems: &Range<T>) -> &Self

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn range_mut(&mut self, elems: &Range<T>) -> &mut Self

pub fn contains(&self, elem: &T) -> bool

Returns true iff set self contains elem.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn words(&self) -> &[Word]

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub fn words_mut(&mut self) -> &mut [Word]

pub fn clone_from(&mut self, other: &IdxSet<T>)

pub fn union(&mut self, other: &IdxSet<T>) -> bool

pub fn subtract(&mut self, other: &IdxSet<T>) -> bool

pub fn intersect(&mut self, other: &IdxSet<T>) -> bool

Important traits for Iter<'a, T>
impl<'a, T: Idx> Iterator for Iter<'a, T> type Item = T;

pub fn iter(&self) -> Iter<T>

pub fn each_bit<F>(&self, max_bits: usize, f: F) where
F: FnMut(T),

Calls f on each index value held in this set, up to the bound max_bits on the size of universe of indexes.

pub fn reset_to_empty(&mut self)

Removes all elements from this set.

Important traits for Elems<'a, T>
impl<'a, T: Idx> Iterator for Elems<'a, T> type Item = T;

pub fn elems(&self, universe_size: usize) -> Elems<T>