Struct rustc_data_structures::indexed_set::IdxSetBuf

pub struct IdxSetBuf<T: Idx> { /* fields omitted */ }

Represents a set (or packed family of sets), of some element type E, where each E is identified by some unique index type T.

In other words, T is the type used to index into the bitvector this type uses to represent the set of object it holds.

Methods

impl<T: Idx> IdxSetBuf<T>

pub fn new_filled(universe_size: usize) -> Self

Creates set holding every element whose index falls in range 0..universe_size.

pub fn new_empty(universe_size: usize) -> Self

Creates set holding no elements.