Struct rustc_data_structures::indexed_vec::IndexVec

pub struct IndexVec<I: Idx, T> {
    pub raw: Vec<T>,
    // some fields omitted
}

Fields
raw: Vec<T>
Methods

impl<I: Idx, T> IndexVec<I, T>

pub fn new() -> Self

pub fn from_raw(raw: Vec<T>) -> Self

pub fn with_capacity(capacity: usize) -> Self

pub fn from_elem<S>(elem: T, universe: &IndexVec<I, S>) -> Self where
T: Clone,

pub fn from_elem_n(elem: T, n: usize) -> Self where
T: Clone,

pub fn push(&mut self, d: T) -> I

pub fn pop(&mut self) -> Option<T>

pub fn len(&self) -> usize

pub fn next_index(&self) -> I

Gives the next index that will be assigned when push is called.

pub fn is_empty(&self) -> bool

Important traits for IntoIter<T>
impl<T> Iterator for IntoIter<T> type Item = T;

pub fn into_iter(self) -> IntoIter<T>

pub fn into_iter_enumerated(self) -> Enumerated<I, IntoIter<T>>

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn iter(&self) -> Iter<T>

pub fn iter_enumerated(&self) -> Enumerated<I, Iter<T>>

Important traits for Map<I, F>
impl<B, I, F> Iterator for Map<I, F> where
F: FnMut(<I as Iterator>::Item) -> B,
I: Iterator, type Item = B;

pub fn indices(&self) -> Map<Range<usize>, IntoIdx<I>>

Important traits for IterMut<'a, T>
impl<'a, T> Iterator for IterMut<'a, T> type Item = &'a mut T;

pub fn iter_mut(&mut self) -> IterMut<T>

pub fn iter_enumerated_mut(&mut self) -> Enumerated<I, IterMut<T>>

pub fn drain<'a, R: RangeBounds<usize>>(
&'a mut self,
range: R
) -> impl Iterator<Item = T> + 'a

pub fn drain_enumerated<'a, R: RangeBounds<usize>>(
&'a mut self,
range: R
) -> impl Iterator<Item = (I, T)> + 'a

pub fn last(&self) -> Option<I>

pub fn shrink_to_fit(&mut self)

pub fn swap(&mut self, a: I, b: I)

pub fn truncate(&mut self, a: usize)

pub fn get(&self, index: I) -> Option<&T>

pub fn get_mut(&mut self, index: I) -> Option<&mut T>

pub fn pick2_mut(&mut self, a: I, b: I) -> (&mut T, &mut T)

Returns mutable references to two distinct elements, a and b. Panics if a == b.

pub fn convert_index_type<Ix: Idx>(self) -> IndexVec<Ix, T>

impl<I: Idx, T: Clone> IndexVec<I, T>

pub fn ensure_contains_elem(&mut self, elem: I, fill_value: impl FnMut() -> T)

Grows the index vector so that it contains an entry for elem; if that is already true, then has no effect. Otherwise, inserts new values as needed by invoking fill_value.

pub fn resize(&mut self, new_len: usize, value: T)

pub fn resize_to_elem(&mut self, elem: I, fill_value: impl FnMut() -> T)

impl<I: Idx, T: Ord> IndexVec<I, T>

pub fn binary_search(&self, value: &T) -> Result<I, I>