Struct rustc_data_structures::indexed_vec::IntoIdx

pub struct IntoIdx<I: Idx> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?