rustc_data_structures::intern
Struct Interned

pub struct Interned<'a, T>(pub &'a T, pub PrivateZst);

A reference to a value that is interned, and is known to be unique.

Note that it is possible to have a T and a Interned<T> that are (or refer to) equal but different values. But if you have two different Interned<T>s, they both refer to the same value, at a single location in memory. This means that equality and hashing can be done on the valueโs address rather than the valueโs contents, which can improve performance.

The PrivateZst field means you can pattern match with Interned(v, _) but you can only construct a Interned with new_unchecked, and not directly.

Tuple Fields
0: &'a T1: PrivateZst
Implementations

impl<'a, T> Interned<'a, T>

pub const fn new_unchecked(t: &'a T) -> Self

Create a new Interned value. The value referred to must be interned and thus be unique, and it must remain unique in the future. This function has _unchecked in the name but is not unsafe, because if the uniqueness condition is violated condition it will cause incorrect behaviour but will not affect memory safety.