Struct rustc_data_structures::ivar::Ivar

pub struct Ivar<T: Copy> {
    // some fields omitted
}

Unstable (rustc_private)

A write-once variable. When constructed, it is empty, and can only be set once.

Ivars ensure that data that can only be initialised once. A full implementation is used for concurrency and blocks on a read of an unfulfilled value. This implementation is more minimal and panics if you attempt to read the value before it has been set. It is also not Sync, but may be extended in the future to be usable as a true concurrency type.

The T: Copy bound is not strictly needed, but it is required by Cell (so removing it would require using UnsafeCell), and it suffices for the current purposes.

Methods

impl<T: Copy> Ivar<T>

fn new() -> Ivar<T>

Unstable (rustc_private)

fn get(&self) -> Option<T>

Unstable (rustc_private)

fn fulfill(&self, value: T)

Unstable (rustc_private)

fn is_fulfilled(&self) -> bool

Unstable (rustc_private)

fn unwrap(&self) -> T

Unstable (rustc_private)