rustc_data_structures

Module jobserver

Structs

Acquired

An acquired token from a jobserver.

Client

A client of a jobserver

HelperThread

Structure returned from Client::into_helper_thread to manage the lifetime of the helper thread returned, see those associated docs for more info.

Proxy

This is a jobserver proxy used to ensure that we hold on to at least one token.

ProxyData 🔒

Constants

ACCESS_ERROR 🔒

Statics

GLOBAL_CLIENT 🔒

GLOBAL_CLIENT_CHECKED 🔒

Functions

client

default_client 🔒

initialize_checked