Struct rustc_data_structures::obligation_forest::Outcome

pub struct Outcome<O, E> {
    pub completed: Vec<O>,
    pub errors: Vec<Error<O, E>>,
    pub stalled: bool,
}

Unstable (rustc_private)

Fields

completed
Unstable (rustc_private)

Obligations that were completely evaluated, including all (transitive) subobligations.

errors
Unstable (rustc_private)

Backtrace of obligations that were found to be in error.

stalled
Unstable (rustc_private)

If true, then we saw no successful obligations, which means there is no point in further iteration. This is based on the assumption that when trait matching returns Err or Ok(None), those results do not affect environmental inference state. (Note that if we invoke process_obligations with no pending obligations, stalled will be true.)