Trait rustc_data_structures::obligation_forest::ForestObligation

pub trait ForestObligation: Clone + Debug {
    type CacheKey: Clone + Hash + Eq + Debug;

    fn as_cache_key(&self) -> Self::CacheKey;
}

Required Associated Types

type CacheKey: Clone + Hash + Eq + Debug

Required Methods

fn as_cache_key(&self) -> Self::CacheKey

Converts this ForestObligation suitable for use as a cache key. If two distinct ForestObligationss return the same cache key, then it must be sound to use the result of processing one obligation (e.g. success for error) for the other obligation