Trait rustc_data_structures::owning_ref::IntoErasedSend

pub unsafe trait IntoErasedSend<'a> {
    type Erased: Send;

    fn into_erased_send(self) -> Self::Erased;
}

Helper trait for erasing the concrete type of what an owner dereferences to, for example Box<T> -> Box<Erased + Send>. This would be unneeded with higher kinded types support in the language.

Required Associated Types

type Erased: Send

Owner with the dereference type substituted to Erased + Send.

Required Methods

fn into_erased_send(self) -> Self::Erased

Performs the type erasure.