Trait rustc_data_structures::owning_ref::IntoErasedSendSync

pub unsafe trait IntoErasedSendSync<'a> {
    type Erased: Send + Sync;

    fn into_erased_send_sync(self) -> Self::Erased;
}

Helper trait for erasing the concrete type of what an owner dereferences to, for example Box<T> -> Box<Erased + Send + Sync>. This would be unneeded with higher kinded types support in the language.

Required Associated Types

type Erased: Send + Sync

Owner with the dereference type substituted to Erased + Send + Sync.

Required Methods

fn into_erased_send_sync(self) -> Self::Erased

Performs the type erasure.