Module rustc_data_structures::sharded

Structs

CacheAligned 🔒

Sharded

An array of cache-line aligned inner locked structures with convenience methods.

Constants

SHARDS

SHARD_BITS 🔒

Traits

IntoPointer

Functions

get_shard_index_by_hash

Get a shard with a pre-computed hash value. If get_shard_by_value is ever used in combination with get_shard_by_hash on a single Sharded instance, then hash must be computed with FxHasher. Otherwise, hash can be computed with any hasher, so long as that hasher is used consistently for each Sharded instance.

make_hash

Type Definitions

ShardedHashMap