Enum rustc_data_structures::snapshot_vec::UndoLog

pub enum UndoLog<D: SnapshotVecDelegate> {
    OpenSnapshot,
    CommittedSnapshot,
    NewElem(usize),
    SetElem(usize, D::Value),
    Other(D::Undo),
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variants
OpenSnapshot

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Indicates where a snapshot started.

CommittedSnapshot

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Indicates a snapshot that has been committed.

NewElem(usize)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

New variable with given index was created.

SetElem(usize, D::Value)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Variable with given index was changed from the given value.

Other(D::Undo)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Extensible set of actions