Trait rustc_data_structures::snapshot_vec::SnapshotVecDelegate

pub trait SnapshotVecDelegate {
    type Value;
    type Undo;
    fn reverse(values: &mut Vec<Self::Value>, action: Self::Undo);
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Associated Types

type Value

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

type Undo

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn reverse(values: &mut Vec<Self::Value>, action: Self::Undo)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?