Module rustc_data_structures::sorted_map

Modules

index_map 🔒

A variant of SortedMap that preserves insertion order.

Structs

SortedIndexMultiMap

An indexed multi-map that preserves insertion order while permitting both O(log n) lookup of an item by key and O(1) lookup by index.

SortedMap

SortedMap is a data structure with similar characteristics as BTreeMap but slightly different trade-offs: lookup, insertion, and removal are O(log(n)) and elements can be iterated in order cheaply.