Struct rustc_data_structures::sorted_map::index_map::SortedIndexMultiMap

pub struct SortedIndexMultiMap<I: Idx, K, V> {
    items: IndexVec<I, (K, V)>,
    idx_sorted_by_item_key: Vec<I>,
}

An indexed multi-map that preserves insertion order while permitting both O(log n) lookup of an item by key and O(1) lookup by index.

This data structure is a hybrid of an IndexVec and a SortedMap. Like IndexVec, SortedIndexMultiMap assigns a typed index to each item while preserving insertion order. Like SortedMap, SortedIndexMultiMap has efficient lookup of items by key. However, this is accomplished by sorting an array of item indices instead of the items themselves.

Unlike SortedMap, this data structure can hold multiple equivalent items at once, so the get_by_key method and its variants return an iterator instead of an Option. Equivalent items will be yielded in insertion order.

Unlike a general-purpose map like BTreeSet or HashSet, SortedMap and SortedIndexMultiMap require O(n) time to insert a single item. This is because we may need to insert into the middle of the sorted array. Users should avoid mutating this data structure in-place.

Fields
items: IndexVec<I, (K, V)>

The elements of the map in insertion order.

idx_sorted_by_item_key: Vec<I>

Indices of the items in the set, sorted by the item’s key.

Implementations

impl<I: Idx, K: Ord, V> SortedIndexMultiMap<I, K, V>

pub fn new() -> Self

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn into_iter(self) -> impl DoubleEndedIterator<Item = (K, V)>

Returns an iterator over the items in the map in insertion order.

pub fn into_iter_enumerated(
self
) -> impl DoubleEndedIterator<Item = (I, (K, V))>

Returns an iterator over the items in the map in insertion order along with their indices.

pub fn iter(&self) -> impl '_ + DoubleEndedIterator<Item = (&K, &V)>

Returns an iterator over the items in the map in insertion order.

pub fn iter_enumerated(
&self
) -> impl '_ + DoubleEndedIterator<Item = (I, (&K, &V))>

Returns an iterator over the items in the map in insertion order along with their indices.

pub fn get(&self, idx: I) -> Option<&(K, V)>

Returns the item in the map with the given index.

pub fn get_by_key(&self, key: K) -> impl Iterator<Item = &V> + '_

Returns an iterator over the items in the map that are equal to key.

If there are multiple items that are equivalent to key, they will be yielded in insertion order.

pub fn get_by_key_enumerated(&self, key: K) -> impl Iterator<Item = (I, &V)> + '_

Returns an iterator over the items in the map that are equal to key along with their indices.

If there are multiple items that are equivalent to key, they will be yielded in insertion order.