Struct rustc_data_structures::sorted_map::SortedMap

pub struct SortedMap<K: Ord, V> { /* fields omitted */ }

SortedMap is a data structure with similar characteristics as BTreeMap but slightly different trade-offs: lookup, insertion, and removal are O(log(N)) and elements can be iterated in order cheaply.

SortedMap can be faster than a BTreeMap for small sizes (<50) since it stores data in a more compact way. It also supports accessing contiguous ranges of elements as a slice, and slices of already sorted elements can be inserted efficiently.

Methods

impl<K: Ord, V> SortedMap<K, V>

pub fn new() -> SortedMap<K, V>

pub fn from_presorted_elements(elements: Vec<(K, V)>) -> SortedMap<K, V>

Construct a SortedMap from a presorted set of elements. This is faster than creating an empty map and then inserting the elements individually.

It is up to the caller to make sure that the elements are sorted by key and that there are no duplicates.

pub fn insert(&mut self, key: K, value: V) -> Option<V>

pub fn remove(&mut self, key: &K) -> Option<V>

pub fn get<Q: ?Sized>(&self, key: &Q) -> Option<&V> where
K: Borrow<Q>,
Q: Ord,

pub fn get_mut<Q: ?Sized>(&mut self, key: &Q) -> Option<&mut V> where
K: Borrow<Q>,
Q: Ord,

pub fn clear(&mut self)

Important traits for Iter<'a, T>
impl<'a, T> Iterator for Iter<'a, T> type Item = &'a T;

pub fn iter(&self) -> Iter<(K, V)>

Iterate over elements, sorted by key

pub fn keys(&self) -> impl Iterator<Item = &K> + ExactSizeIterator

Iterate over the keys, sorted

pub fn values(&self) -> impl Iterator<Item = &V> + ExactSizeIterator

Iterate over values, sorted by key

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

Important traits for &'_ [u8]
impl<'_> Read for &'_ [u8]impl<'_> Write for &'_ mut [u8]

pub fn range<R>(&self, range: R) -> &[(K, V)] where
R: RangeBounds<K>,

pub fn remove_range<R>(&mut self, range: R) where
R: RangeBounds<K>,

pub fn offset_keys<F>(&mut self, f: F) where
F: Fn(&mut K),

Mutate all keys with the given function f. This mutation must not change the sort-order of keys.

pub fn insert_presorted(&mut self, elements: Vec<(K, V)>)

Inserts a presorted range of elements into the map. If the range can be inserted as a whole in between to existing elements of the map, this will be faster than inserting the elements individually.

It is up to the caller to make sure that the elements are sorted by key and that there are no duplicates.

pub fn contains_key<Q: ?Sized>(&self, key: &Q) -> bool where
K: Borrow<Q>,
Q: Ord,