rustc_data_structures::sso

Module map

Structs

Entry

A view into a single entry in a map.

Enums

SsoHashMap

Small-storage-optimized implementation of a map.

Constants

SSO_ARRAY_SIZE 🔒

For pointer-sized arguments arrays are faster than set/map for up to 64 arguments.

Functions

adapt_array_mut_it 🔒

adapts Item of array mut reference iterator to Item of hashmap mut reference iterator.

adapt_array_ref_it 🔒

adapts Item of array reference iterator to Item of hashmap reference iterator.