Struct rustc_data_structures::sso::map::Entry

pub struct Entry<'a, K, V> {
    ssomap: &'a mut SsoHashMap<K, V>,
    key: K,
}

A view into a single entry in a map.

Fields
ssomap: &'a mut SsoHashMap<K, V>key: K
Implementations

impl<'a, K: Eq + Hash, V> Entry<'a, K, V>

pub fn and_modify<F>(self, f: F) -> Selfwhere
F: FnOnce(&mut V),

Provides in-place mutable access to an occupied entry before any potential inserts into the map.

pub fn or_insert(self, value: V) -> &'a mut V

Ensures a value is in the entry by inserting the default if empty, and returns a mutable reference to the value in the entry.

pub fn or_insert_with<F: FnOnce() -> V>(self, default: F) -> &'a mut V

Ensures a value is in the entry by inserting the result of the default function if empty, and returns a mutable reference to the value in the entry.

pub fn key(&self) -> &K

Returns a reference to this entry’s key.

impl<'a, K: Eq + Hash, V: Default> Entry<'a, K, V>

pub fn or_default(self) -> &'a mut V

Ensures a value is in the entry by inserting the default value if empty, and returns a mutable reference to the value in the entry.