Struct rustc_data_structures::stable_hasher::StableVec

pub struct StableVec<T>(_);

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A vector container that makes sure that its items are hashed in a stable order.

Methods

impl<T> StableVec<T>

pub fn new(v: Vec<T>) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?