Trait rustc_data_structures::stable_hasher::StableHashingContextProvider

pub trait StableHashingContextProvider {
    type ContextType;
    fn create_stable_hashing_context(&self) -> Self::ContextType;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Something that can provide a stable hashing context.

Associated Types

type ContextType

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Required Methods

fn create_stable_hashing_context(&self) -> Self::ContextType

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?