Module rustc_data_structures::sync::parallel

This module defines parallel operations that are implemented in one way for the serial compiler, and another way the parallel compiler.

Re-exports

pub use disabled::*;

Modules

disabled 🔒

Structs

ParallelGuard

A guard used to hold panics that occur during a parallel section to later by unwound. This is used for the parallel compiler to prevent fatal errors from non-deterministically hiding errors by ensuring that everything in the section has completed executing before continuing with unwinding. It’s also used for the non-parallel code to ensure error message output match the parallel compiler for testing purposes.

Functions

parallel_guard

This gives access to a fresh parallel guard in the closure and will unwind any panics caught in it after the closure returns.