Struct rustc_data_structures::sync::Weak

pub struct Weak<T> where
    T: ?Sized,  { /* fields omitted */ }

Weak is a version of Rc that holds a non-owning reference to the managed value. The value is accessed by calling upgrade on the Weak pointer, which returns an Option<Rc<T>>.

Since a Weak reference does not count towards ownership, it will not prevent the inner value from being dropped, and Weak itself makes no guarantees about the value still being present and may return None when upgraded.

A Weak pointer is useful for keeping a temporary reference to the value within Rc without extending its lifetime. It is also used to prevent circular references between Rc pointers, since mutual owning references would never allow either Rc to be dropped. For example, a tree could have strong Rc pointers from parent nodes to children, and Weak pointers from children back to their parents.

The typical way to obtain a Weak pointer is to call Rc::downgrade.

Methods

impl<T> Weak<T>

pub fn new() -> Weak<T>

Constructs a new Weak<T>, without allocating any memory. Calling upgrade on the return value always gives None.

Examples

use std::rc::Weak;

let empty: Weak<i64> = Weak::new();
assert!(empty.upgrade().is_none());

impl<T> Weak<T> where
T: ?Sized,

pub fn upgrade(&self) -> Option<Rc<T>>

Attempts to upgrade the Weak pointer to an Rc, extending the lifetime of the value if successful.

Returns None if the value has since been dropped.

Examples

use std::rc::Rc;

let five = Rc::new(5);

let weak_five = Rc::downgrade(&five);

let strong_five: Option<Rc<_>> = weak_five.upgrade();
assert!(strong_five.is_some());

// Destroy all strong pointers.
drop(strong_five);
drop(five);

assert!(weak_five.upgrade().is_none());

pub fn strong_count(&self) -> usize

🔬 This is a nightly-only experimental API. (weak_counts)

Gets the number of strong (Rc) pointers pointing to this value.

If self was created using Weak::new, this will return 0.

pub fn weak_count(&self) -> Option<usize>

🔬 This is a nightly-only experimental API. (weak_counts)

Gets the number of Weak pointers pointing to this value.

If self was created using Weak::new, this will return None. If not, the returned value is at least 1, since self still points to the value.

pub fn ptr_eq(this: &Weak<T>, other: &Weak<T>) -> bool

🔬 This is a nightly-only experimental API. (weak_ptr_eq)

Returns true if the two Weaks point to the same value (not just values that compare as equal).

Notes

Since this compares pointers it means that Weak::new() will equal each other, even though they don't point to any value.

Examples

#![feature(weak_ptr_eq)]
use std::rc::{Rc, Weak};

let first_rc = Rc::new(5);
let first = Rc::downgrade(&first_rc);
let second = Rc::downgrade(&first_rc);

assert!(Weak::ptr_eq(&first, &second));

let third_rc = Rc::new(5);
let third = Rc::downgrade(&third_rc);

assert!(!Weak::ptr_eq(&first, &third));

Comparing Weak::new.

#![feature(weak_ptr_eq)]
use std::rc::{Rc, Weak};

let first = Weak::new();
let second = Weak::new();
assert!(Weak::ptr_eq(&first, &second));

let third_rc = Rc::new(());
let third = Rc::downgrade(&third_rc);
assert!(!Weak::ptr_eq(&first, &third));