Struct rustc_data_structures::sync::WriteGuard

pub struct WriteGuard<'b, T> where
    T: 'b + ?Sized,  { /* fields omitted */ }

A wrapper type for a mutably borrowed value from a RefCell<T>.

See the module-level documentation for more.

Methods

impl<'b, T> RefMut<'b, T> where
T: ?Sized,

pub fn map<U, F>(orig: RefMut<'b, T>, f: F) -> RefMut<'b, U> where
F: FnOnce(&mut T) -> &mut U,
U: ?Sized,

Make a new RefMut for a component of the borrowed data, e.g. an enum variant.

The RefCell is already mutably borrowed, so this cannot fail.

This is an associated function that needs to be used as RefMut::map(...). A method would interfere with methods of the same name on the contents of a RefCell used through Deref.

Examples

use std::cell::{RefCell, RefMut};

let c = RefCell::new((5, 'b'));
{
    let b1: RefMut<(u32, char)> = c.borrow_mut();
    let mut b2: RefMut<u32> = RefMut::map(b1, |t| &mut t.0);
    assert_eq!(*b2, 5);
    *b2 = 42;
}
assert_eq!(*c.borrow(), (42, 'b'));