rustc_data_structures::sync
Type Alias MappedWriteGuard

pub type MappedWriteGuard<'a, T> = MappedRwLockWriteGuard<'a, RawRwLock, T>;

An RAII write lock guard returned by RwLockWriteGuard::map, which can point to a subfield of the protected data.

The main difference between MappedRwLockWriteGuard and RwLockWriteGuard is that the former doesn’t support temporarily unlocking and re-locking, since that could introduce soundness issues if the locked object is modified by another thread.

Aliased Type

struct MappedWriteGuard<'a, T> { /* private fields */ }

Layout

Note: Most layout information is completely unstable and may even differ between compilations. The only exception is types with certain repr(...) attributes. Please see the Rust Reference's “Type Layout” chapter for details on type layout guarantees.

Size: 16 bytes