Struct rustc_data_structures::sync::worker_local::Registry

pub struct Registry(Arc<RegistryData>);

Represents a list of threads which can access worker locals.

Tuple Fields
0: Arc<RegistryData>
Implementations

impl Registry

pub fn new(thread_limit: usize) -> Self

Creates a registry which can hold up to thread_limit threads.

pub fn current() -> Self

Gets the registry associated with the current thread. Panics if there’s no such registry.

pub fn register(&self)

Registers the current thread with the registry so worker locals can be used on it. Panics if the thread limit is hit or if the thread already has an associated registry.

fn id(&self) -> RegistryId

Gets the identifer of this registry.