Struct rustc_data_structures::tagged_ptr::copy::CopyTaggedPtr

pub struct CopyTaggedPtr<P, T, const COMPARE_PACKED: bool>
where
    P: Pointer,
    T: Tag,
{
    packed: NonNull<P::Target>,
    tag_ghost: PhantomData<T>,
}

A Copy tagged pointer.

This is essentially { pointer: P, tag: T } packed in a single pointer.

You should use this instead of the TaggedPtr type in all cases where P implements Copy.

If COMPARE_PACKED is true, then the pointers will be compared and hashed without unpacking. Otherwise we don’t implement PartialEq, Eq and Hash; if you want that, wrap the CopyTaggedPtr.

Fields
packed: NonNull<P::Target>

This is semantically a pair of pointer: P and tag: T fields, however we pack them in a single pointer, to save space.

We pack the tag into the most-significant bits of the pointer to ease retrieval of the value. A left shift is a multiplication and those are embeddable in instruction encoding, for example:

// (<https://godbolt.org/z/jqcYPWEr3>)
example::shift_read3:
    mov     eax, dword ptr [8*rdi]
    ret

example::mask_read3:
    and     rdi, -8
    mov     eax, dword ptr [rdi]
    ret

This is ASM outputted by rustc for reads of values behind tagged pointers for different approaches of tagging:

shift_read3 uses << 3 (the tag is in the most-significant bits)

mask_read3 uses & !0b111 (the tag is in the least-significant bits)

The shift approach thus produces less instructions and is likely faster (see https://godbolt.org/z/Y913sMdWb).

Encoding diagram:

[ packed.addr                     ]
[ tag ] [ pointer.addr >> T::BITS ] <-- usize::BITS - T::BITS bits
   ^
   |
T::BITS bits

The tag can be retrieved by packed.addr() >> T::BITS and the pointer can be retrieved by packed.map_addr(|addr| addr << T::BITS).

tag_ghost: PhantomData<T>
Implementations

impl<P, T, const CP: bool> CopyTaggedPtr<P, T, CP>
where P: Pointer, T: Tag,

pub fn new(pointer: P, tag: T) -> Self

Tags pointer with tag.

Note that this leaks pointer: it won’t be dropped when CopyTaggedPtr is dropped. If you have a pointer with a significant drop, use TaggedPtr instead.

pub fn pointer(self) -> P
where P: Copy,

Retrieves the pointer.

pub fn tag(&self) -> T

Retrieves the tag.

pub fn set_tag(&mut self, tag: T)

Sets the tag to a new value.

const TAG_BIT_SHIFT: u32 = _

const ASSERTION: () = _

fn pack(ptr: NonNull<P::Target>, tag: T) -> NonNull<P::Target>

Pack pointer ptr that comes from P::into_ptr with a tag, according to self.packed encoding scheme.

pub(super) fn pointer_raw(&self) -> NonNull<P::Target>

Retrieves the original raw pointer from self.packed.

pub(super) fn with_pointer_ref<R>(&self, f: impl FnOnce(&P) -> R) -> R

This provides a reference to the P pointer itself, rather than the Deref::Target. It is used for cases where we want to call methods that may be implement differently for the Pointer than the Pointee (e.g., Rc::clone vs cloning the inner value).