Struct rustc_data_structures::tagged_ptr::drop::TaggedPtr

pub struct TaggedPtr<P, T, const COMPARE_PACKED: bool>
where
    P: Pointer,
    T: Tag,
{
    raw: CopyTaggedPtr<P, T, COMPARE_PACKED>,
}

A tagged pointer that supports pointers that implement Drop.

This is essentially { pointer: P, tag: T } packed in a single pointer.

You should use CopyTaggedPtr instead of the this type in all cases where P implements Copy.

If COMPARE_PACKED is true, then the pointers will be compared and hashed without unpacking. Otherwise we donโt implement PartialEq, Eq and Hash; if you want that, wrap the TaggedPtr.

Fields
raw: CopyTaggedPtr<P, T, COMPARE_PACKED>
Implementations

impl<P, T, const CP: bool> TaggedPtr<P, T, CP>
where P: Pointer, T: Tag,

pub fn new(pointer: P, tag: T) -> Self

Tags pointer with tag.

pub fn tag(&self) -> T

Retrieves the tag.

pub fn set_tag(&mut self, tag: T)

Sets the tag to a new value.