Trait rustc_data_structures::ToHex

pub trait ToHex {
    fn to_hex(&self) -> String;
}

A trait for converting a value to hexadecimal encoding

Required methods

fn to_hex(&self) -> String

Converts the value of self to a hex value, returning the owned string.

Loading content...