Trait rustc_data_structures::tuple_slice::TupleSlice

pub trait TupleSlice<T> {
    fn as_slice(&self) -> &[T];

    fn as_mut_slice(&mut self) -> &mut [T];
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Allows to view uniform tuples as slices

Required Methods

fn as_slice(&self) -> &[T]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

fn as_mut_slice(&mut self) -> &mut [T]

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?