Struct rustc_data_structures::unify::UnificationTable

pub struct UnificationTable<K: UnifyKey> {
    values: SnapshotVec<Delegate<K>>,
}

Table of unification keys and their values.

Fields
values: SnapshotVec<Delegate<K>>

Indicates the current value of each key.

Methods

impl<K: UnifyKey> UnificationTable<K>

We can't use V:LatticeValue, much as I would like to, because frequently the pattern is that V=Option for some other type parameter U, and we have no way to say Option:LatticeValue.

pub fn new() -> UnificationTable<K>

pub fn snapshot(&mut self) -> Snapshot<K>

Starts a new snapshot. Each snapshot must be either rolled back or committed in a "LIFO" (stack) order.

pub fn rollback_to(&mut self, snapshot: Snapshot<K>)

Reverses all changes since the last snapshot. Also removes any keys that have been created since then.

pub fn commit(&mut self, snapshot: Snapshot<K>)

Commits all changes since the last snapshot. Of course, they can still be undone if there is a snapshot further out.

pub fn new_key(&mut self, value: K::Value) -> K

fn get(&mut self, vid: K) -> VarValue<K>

Find the root node for vid. This uses the standard union-find algorithm with path compression: http://en.wikipedia.org/wiki/Disjoint-set_data_structure.

NB. This is a building-block operation and you would probably prefer to call probe below.

fn is_root(&self, key: K) -> bool

fn set(&mut self, key: K, new_value: VarValue<K>)

Sets the value for vid to new_value. vid MUST be a root node! This is an internal operation used to impl other things.

fn unify(
&mut self,
root_a: VarValue<K>,
root_b: VarValue<K>,
new_value: K::Value
) -> K

Either redirects node_a to node_b or vice versa, depending on the relative rank. The value associated with the new root will be new_value.

NB: This is the "union" operation of "union-find". It is really more of a building block. If the values associated with your key are non-trivial, you would probably prefer to call unify_var_var below.

fn redirect_root(
&mut self,
new_rank: u32,
old_root: VarValue<K>,
new_root: VarValue<K>,
new_value: K::Value
) -> K

impl<'tcx, K: UnifyKey> UnificationTable<K> where
K::Value: Combine,

Base union-find algorithm, where we are just making sets

pub fn union(&mut self, a_id: K, b_id: K) -> K

pub fn find(&mut self, id: K) -> K

pub fn find_value(&mut self, id: K) -> K::Value

impl<'tcx, K, V> UnificationTable<K> where
K: UnifyKey<Value = Option<V>>,
V: Clone + PartialEq + Debug,

Non-subtyping unification

Code to handle keys which carry a value, like ints, floats---anything that doesn't have a subtyping relationship we need to worry about.

pub fn unify_var_var(&mut self, a_id: K, b_id: K) -> Result<K, (V, V)>

pub fn unify_var_value(&mut self, a_id: K, b: V) -> Result<(), (V, V)>

Sets the value of the key a_id to b. Because simple keys do not have any subtyping relationships, if a_id already has a value, it must be the same as b.

pub fn has_value(&mut self, id: K) -> bool

pub fn probe(&mut self, a_id: K) -> Option<V>

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn unsolved_variables(&mut self) -> Vec<K>