Struct rustc_data_structures::unify::VarValue

pub struct VarValue<K: UnifyKey> {
    parent: K,
    value: K::Value,
    rank: u32,
}

Value of a unification key. We implement Tarjan's union-find algorithm: when two keys are unified, one of them is converted into a "redirect" pointing at the other. These redirects form a DAG: the roots of the DAG (nodes that are not redirected) are each associated with a value of type V and a rank. The rank is used to keep the DAG relatively balanced, which helps keep the running time of the algorithm under control. For more information, see http://en.wikipedia.org/wiki/Disjoint-set_data_structure.

Fields
parent: K value: K::Value rank: u32
Methods

impl<K: UnifyKey> VarValue<K>

fn new_var(key: K, value: K::Value) -> VarValue<K>

fn new(parent: K, value: K::Value, rank: u32) -> VarValue<K>

fn redirect(self, to: K) -> VarValue<K>

fn root(self, rank: u32, value: K::Value) -> VarValue<K>

fn key(&self) -> K

Returns the key of this node. Only valid if this is a root node, which you yourself must ensure.

fn parent(&self, self_key: K) -> Option<K>

fn if_not_self(&self, key: K, self_key: K) -> Option<K>