Trait rustc_data_structures::unify::Combine

pub trait Combine {
    fn combine(&self, other: &Self) -> Self;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

This trait is implemented for unify values that can be combined. This relation should be a monoid.

Required Methods

fn combine(&self, other: &Self) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?