Struct rustc_data_structures::unord::UnordBag

pub struct UnordBag<V> {
    inner: Vec<V>,
}

This is a collection type that tries very hard to not expose any internal iteration. This is a useful property when trying to uphold the determinism invariants imposed by the query system.

This collection type is a good choice for collections the keys of which donโt have a semantic ordering and donโt implement Hash or Eq.

See MCP 533 for more information.

Fields
inner: Vec<V>
Implementations

impl<V> UnordBag<V>

pub fn new() -> Self

pub fn len(&self) -> usize

pub fn push(&mut self, v: V)

pub fn items(&self) -> UnordItems<&V, impl Iterator<Item = &V>>

pub fn into_items(self) -> UnordItems<V, impl Iterator<Item = V>>