rustc_data_structures::unord
Struct UnordItems

pub struct UnordItems<T, I: Iterator<Item = T>>(I);

UnordItems is the order-less version of Iterator. It only contains methods that donโt (easily) expose an ordering of the underlying items.

Most methods take an Fn where the Iterator-version takes an FnMut. This is to reduce the risk of accidentally leaking the internal order via the closure environment. Otherwise one could easily do something like

let mut ordered = vec![];
unordered_items.all(|x| ordered.push(x));

Itโs still possible to do the same thing with an Fn by using interior mutability, but the chance of doing it accidentally is reduced.

Tuple Fields
0: I
Implementations

impl<T, I: Iterator<Item = T>> UnordItems<T, I>

pub fn map<U, F: Fn(T) -> U>( self, f: F, ) -> UnordItems<U, impl Iterator<Item = U>>

pub fn all<F: Fn(T) -> bool>(self, f: F) -> bool

pub fn any<F: Fn(T) -> bool>(self, f: F) -> bool

pub fn filter<F: Fn(&T) -> bool>( self, f: F, ) -> UnordItems<T, impl Iterator<Item = T>>

pub fn filter_map<U, F: Fn(T) -> Option<U>>( self, f: F, ) -> UnordItems<U, impl Iterator<Item = U>>

pub fn max(self) -> Option<T>
where T: Ord,

pub fn min(self) -> Option<T>
where T: Ord,

pub fn sum<S>(self) -> S
where S: Sum<T>,

pub fn product<S>(self) -> S
where S: Product<T>,

pub fn count(self) -> usize

pub fn flat_map<U, F, O>(self, f: F) -> UnordItems<O, impl Iterator<Item = O>>
where U: IntoIterator<Item = O>, F: Fn(T) -> U,

pub fn collect<C: From<UnordItems<T, I>>>(self) -> C

impl<T> UnordItems<T, Empty<T>>

pub fn empty() -> Self

impl<'a, T: Clone + 'a, I: Iterator<Item = &'a T>> UnordItems<&'a T, I>

pub fn cloned(self) -> UnordItems<T, impl Iterator<Item = T>>

impl<'a, T: Copy + 'a, I: Iterator<Item = &'a T>> UnordItems<&'a T, I>

pub fn copied(self) -> UnordItems<T, impl Iterator<Item = T>>

impl<T, I: Iterator<Item = T>> UnordItems<T, I>

pub fn into_sorted<HCX>(self, hcx: &HCX) -> Vec<T>
where T: ToStableHashKey<HCX>,

pub fn into_sorted_stable_ord(self) -> Vec<T>
where T: StableCompare,

pub fn into_sorted_stable_ord_by_key<K, C>(self, project_to_key: C) -> Vec<T>
where K: StableCompare, C: for<'a> Fn(&'a T) -> &'a K,

pub fn collect_sorted<HCX, C>(self, hcx: &HCX, cache_sort_key: bool) -> C
where T: ToStableHashKey<HCX>, C: FromIterator<T> + BorrowMut<[T]>,

pub fn collect_stable_ord_by_key<K, C, P>(self, project_to_key: P) -> C
where K: StableCompare, P: for<'a> Fn(&'a T) -> &'a K, C: FromIterator<T> + BorrowMut<[T]>,