rustc_data_structures::unord
Struct UnordSet

pub struct UnordSet<V: Eq + Hash> {
    inner: FxHashSet<V>,
}

This is a set collection type that tries very hard to not expose any internal iteration. This is a useful property when trying to uphold the determinism invariants imposed by the query system.

This collection type is a good choice for set-like collections the keys of which donโt have a semantic ordering.

See MCP 533 for more information.

Fields
inner: FxHashSet<V>
Implementations

impl<V: Eq + Hash> UnordSet<V>

pub fn new() -> Self

pub fn with_capacity(capacity: usize) -> Self

pub fn len(&self) -> usize

pub fn is_empty(&self) -> bool

pub fn insert(&mut self, v: V) -> bool

pub fn contains<Q>(&self, v: &Q) -> bool
where V: Borrow<Q>, Q: Hash + Eq + ?Sized,

pub fn remove<Q>(&mut self, k: &Q) -> bool
where V: Borrow<Q>, Q: Hash + Eq + ?Sized,

pub fn items(&self) -> UnordItems<&V, impl Iterator<Item = &V>>

pub fn into_items(self) -> UnordItems<V, impl Iterator<Item = V>>

pub fn to_sorted<HCX>(&self, hcx: &HCX, cache_sort_key: bool) -> Vec<&V>
where V: ToStableHashKey<HCX>,

Returns the items of this set in stable sort order (as defined by ToStableHashKey).

The cache_sort_key parameter controls if slice::sort_by_cached_key or slice::sort_unstable_by_key will be used for sorting the vec. Use cache_sort_key when the ToStableHashKey::to_stable_hash_key implementation for V is expensive (e.g. a DefId -> DefPathHash lookup).

pub fn to_sorted_stable_ord(&self) -> Vec<&V>
where V: StableCompare,

Returns the items of this set in stable sort order (as defined by StableCompare). This method is much more efficient than into_sorted because it does not need to transform keys to their ToStableHashKey equivalent.

pub fn into_sorted_stable_ord(self) -> Vec<V>
where V: StableCompare,

Returns the items of this set in stable sort order (as defined by StableCompare). This method is much more efficient than into_sorted because it does not need to transform keys to their ToStableHashKey equivalent.

pub fn into_sorted<HCX>(self, hcx: &HCX, cache_sort_key: bool) -> Vec<V>
where V: ToStableHashKey<HCX>,

Returns the items of this set in stable sort order (as defined by ToStableHashKey).

The cache_sort_key parameter controls if slice::sort_by_cached_key or slice::sort_unstable_by_key will be used for sorting the vec. Use cache_sort_key when the ToStableHashKey::to_stable_hash_key implementation for V is expensive (e.g. a DefId -> DefPathHash lookup).

pub fn clear(&mut self)