Trait rustc_data_structures::unord::ExtendUnord

pub trait ExtendUnord<T> {
    // Required method
    fn extend_unord<I: Iterator<Item = T>>(&mut self, items: UnordItems<T, I>);
}

Required Methods

fn extend_unord<I: Iterator<Item = T>>(&mut self, items: UnordItems<T, I>)

Extend this unord collection with the given UnordItems. This method is called extend_unord instead of just extend so it does not conflict with Extend::extend. Otherwise there would be many places where the two methods would have to be explicitly disambiguated via UFCS.

Object Safety

This trait is not object safe.