Struct rustc_data_structures::vec_map::VecMap

pub struct VecMap<K, V>(Vec<(K, V)>);

A map type implemented as a vector of pairs K (key) and V (value). It currently provides a subset of all the map operations, the rest could be added as needed.

Tuple Fields
0: Vec<(K, V)>
Implementations

impl<K, V> VecMap<K, V>where
K: Debug + PartialEq,
V: Debug,

pub fn new() -> Self

pub fn insert(&mut self, k: K, v: V) -> Option<V>

Sets the value of the entry, and returns the entry’s old value.

pub fn remove(&mut self, k: &K) -> Option<V>

Removes the entry from the map and returns the removed value

pub fn get<Q: ?Sized>(&self, k: &Q) -> Option<&V>where
K: Borrow<Q>,
Q: Eq,

Gets a reference to the value in the entry.

pub fn get_mut<Q: ?Sized>(&mut self, k: &Q) -> Option<&mut V>where
K: Borrow<Q>,
Q: Eq,

Gets a mutable reference to the value in the entry.

pub fn any_value_matching(
&self,
predicate: impl FnMut(&(K, V)) -> bool
) -> Option<&V>

Returns the any value corresponding to the supplied predicate filter.

The supplied predicate will be applied to each (key, value) pair and it will return a reference to the values where the predicate returns true.

pub fn get_value_matching(
&self,
predicate: impl FnMut(&(K, V)) -> bool
) -> Option<&V>

Returns the value corresponding to the supplied predicate filter. It crashes if there’s more than one matching element.

The supplied predicate will be applied to each (key, value) pair and it will return a reference to the value where the predicate returns true.

pub fn contains_key<Q: ?Sized>(&self, k: &Q) -> boolwhere
K: Borrow<Q>,
Q: Eq,

Returns true if the map contains a value for the specified key.

The key may be any borrowed form of the map’s key type, Eq on the borrowed form must match those for the key type.

pub fn is_empty(&self) -> bool

Returns true if the map contains no elements.

pub fn iter(&self) -> Iter<'_, (K, V)>

pub fn iter_mut(&mut self) -> impl Iterator<Item = (&K, &mut V)>

pub fn retain(&mut self, f: impl Fn(&(K, V)) -> bool)