Crate rustc_data_structures

Unstable

Various data structures used by the Rust compiler. The intention is that code in here should be not be specific to rustc, so that it can be easily unit tested and so forth.

Note

This API is completely unstable and subject to change.

Modules

bitvec 	 [Unstable]

graph 	 [Unstable]
A graph module for use in dataflow, region resolution, and elsewhere.

snapshot_vec 	 [Unstable]
A utility class for implementing "snapshottable" things; a snapshottable data structure permits you to take a snapshot (via start_snapshot) and then, after making some changes, elect either to rollback to the start of the snapshot or commit those changes.

unify 	 [Unstable]