Crate rustc_data_structures

Unstable (rustc_private)

Various data structures used by the Rust compiler. The intention is that code in here should be not be specific to rustc, so that it can be easily unit tested and so forth.

Note

This API is completely unstable and subject to change.

Modules

bitvec 	 [Unstable]

fnv 	 [Unstable]

graph 	 [Unstable]
A graph module for use in dataflow, region resolution, and elsewhere.

ivar 	 [Unstable]

obligation_forest 	 [Unstable]
The ObligationForest is a utility data structure used in trait matching to track the set of outstanding obligations (those not yet resolved to success or error). It also tracks the "backtrace" of each pending obligation (why we are trying to figure this out in the first place). See README.md for a general overview of how to use this class.

snapshot_vec 	 [Unstable]
A utility class for implementing "snapshottable" things; a snapshottable data structure permits you to take a snapshot (via start_snapshot) and then, after making some changes, elect either to rollback to the start of the snapshot or commit those changes.

transitive_relation 	 [Unstable]

tuple_slice 	 [Unstable]

unify 	 [Unstable]

veccell 	 [Unstable]