Struct rustc_data_structures::bitvec::BitMatrix

pub struct BitMatrix { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A "bit matrix" is basically a matrix of booleans represented as one gigantic bitvector. In other words, it is as if you have rows bitvectors, each of length columns.

Methods

impl BitMatrix

pub fn new(rows: usize, columns: usize) -> BitMatrix

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn add(&mut self, source: usize, target: usize) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn contains(&self, source: usize, target: usize) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Do the bits from source contain target?

Put another way, if the matrix represents (transitive) reachability, can source reach target?

pub fn intersection(&self, a: usize, b: usize) -> Vec<usize>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns those indices that are reachable from both a and b. This is an O(n) operation where n is the number of elements (somewhat independent from the actual size of the intersection, in particular).

pub fn merge(&mut self, read: usize, write: usize) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Add the bits from read to the bits from write, return true if anything changed.

This is used when computing transitive reachability because if you have an edge write -> read, because in that case write can reach everything that read can (and potentially more).

pub fn iter<'a>(&'a self, row: usize) -> BitVectorIter<'a>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?