rustc_data_structures
Module fx

Structs

FxHasher

A speedy hash algorithm for use within rustc. The hashmap in liballoc by default uses SipHash which isn’t quite as speedy as we want. In the compiler we’re not really worried about DOS attempts, so we use a fast non-cryptographic hash.

Type Aliases

FxHashMap

Type alias for a hashmap using the fx hash algorithm.

FxHashSet

Type alias for a hashmap using the fx hash algorithm.

FxIndexMap

FxIndexSet

IndexEntry

IndexOccupiedEntry

StdEntry