rustc_data_structures
Module graph

Modules

dominators

Finding the dominators in a control-flow graph.

implementation

A graph module for use in dataflow, region resolution, and elsewhere.

iterate

reference 🔒

scc

Routine to compute the strongly connected components (SCCs) of a graph.

vec_graph

Traits

ControlFlowGraph

Alias for DirectedGraph + StartNode + Predecessors + Successors.

DirectedGraph

NumEdges

Predecessors

StartNode

Successors

Functions

depth_first_search

depth_first_search_as_undirected

is_cyclic

Returns true if the graph has a cycle that is reachable from the start node.