rustc_data_structures::graph::vec_graph
Struct VecGraph

pub struct VecGraph<N: Idx, const BR: bool = false> {
    node_starts: IndexVec<N, usize>,
    edge_targets: Vec<N>,
}

A directed graph, efficient for cases where node indices are pre-existing.

If BR is true, the graph will store back-references, allowing you to get predecessors.

Fields
node_starts: IndexVec<N, usize>

Indices into edge_targets that signify a start of list of edges.

edge_targets: Vec<N>

Targets (or sources for back refs) of edges

Implementations

impl<N: Idx + Ord, const BR: bool> VecGraph<N, BR>

pub fn new(num_nodes: usize, edge_pairs: Vec<(N, N)>) -> Self

pub fn successors(&self, source: N) -> &[N]

Gets the successors for source as a slice.

impl<N: Idx + Ord> VecGraph<N, true>

pub fn predecessors(&self, target: N) -> &[N]

Gets the predecessors for target as a slice.