Struct rustc_data_structures::indexed_set::IdxSetBuf

pub struct IdxSetBuf<T: Idx> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Represents a set (or packed family of sets), of some element type E, where each E is identified by some unique index type T.

In other words, T is the type used to index into the bitvector this type uses to represent the set of object it holds.

Methods

impl<T: Idx> IdxSetBuf<T>

pub fn new_filled(universe_size: usize) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Creates set holding every element whose index falls in range 0..universe_size.

pub fn new_empty(universe_size: usize) -> Self

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Creates set holding no elements.