Module rustc_data_structures::obligation_forest

Unstable (rustc_private)

The ObligationForest is a utility data structure used in trait matching to track the set of outstanding obligations (those not yet resolved to success or error). It also tracks the "backtrace" of each pending obligation (why we are trying to figure this out in the first place). See README.md for a general overview of how to use this class.

Structs

Error 	 [Unstable]

ObligationForest 	 [Unstable]

Outcome 	 [Unstable]

Snapshot 	 [Unstable]

Traits

ForestObligation 	 [Unstable]

ObligationProcessor 	 [Unstable]