Struct rustc_data_structures::obligation_forest::ObligationForest

pub struct ObligationForest<O, T> {
    // some fields omitted
}

Unstable (rustc_private)

Methods

impl<O: Debug, T: Debug> ObligationForest<O, T>

fn new() -> ObligationForest<O, T>

Unstable (rustc_private)

fn len(&self) -> usize

Unstable (rustc_private)

Return the total number of nodes in the forest that have not yet been fully resolved.

fn start_snapshot(&mut self) -> Snapshot

Unstable (rustc_private)

fn commit_snapshot(&mut self, snapshot: Snapshot)

Unstable (rustc_private)

fn rollback_snapshot(&mut self, snapshot: Snapshot)

Unstable (rustc_private)

fn in_snapshot(&self) -> bool

Unstable (rustc_private)

fn push_tree(&mut self, obligation: O, tree_state: T)

Unstable (rustc_private)

Adds a new tree to the forest.

This CAN be done during a snapshot.

fn to_errors<E: Clone>(&mut self, error: E) -> Vec<Error<O, E>>

Unstable (rustc_private)

Convert all remaining obligations to the given error.

This cannot be done during a snapshot.

fn pending_obligations(&self) -> Vec<O> where O: Clone

Unstable (rustc_private)

Returns the set of obligations that are in a pending state.

fn process_obligations<E, F>(&mut self, action: F) -> Outcome<O, E> where E: Debug, F: FnMut(&mut O, &mut T, Backtrace<O>) -> Result<Option<Vec<O>>, E>

Unstable (rustc_private)

Process the obligations.

This CANNOT be unrolled (presently, at least).