Struct rustc_data_structures::obligation_forest::ObligationForest

pub struct ObligationForest<O: ForestObligation> { /* fields omitted */ }

Methods

impl<O: ForestObligation> ObligationForest<O>

pub fn dump_graphviz<P: AsRef<Path>>(&self, dir: P, description: &str)

Creates a graphviz representation of the obligation forest. Given a directory this will create files with name of the format <counter>_<description>.gv. The counter is global and is maintained internally.

Calling this will do nothing unless the environment variable DUMP_OBLIGATION_FOREST_GRAPHVIZ is defined.

A few post-processing that you might want to do make the forest easier to visualize:

sed 's,std::[a-z]*::,,g' — Deletes the std::<package>:: prefix of paths.

sed 's,"Binder(TraitPredicate(<\(.*\)>)) (\([^)]*\))","\1 (\2)",' — Transforms Binder(TraitPredicate(<predicate>)) into just <predicate>.

impl<O: ForestObligation> ObligationForest<O>

pub fn new() -> ObligationForest<O>

pub fn len(&self) -> usize

Returns the total number of nodes in the forest that have not yet been fully resolved.

pub fn register_obligation(&mut self, obligation: O)

Registers an obligation.

This CAN be done in a snapshot

pub fn to_errors<E: Clone>(&mut self, error: E) -> Vec<Error<O, E>>

Converts all remaining obligations to the given error.

This cannot be done during a snapshot.

pub fn map_pending_obligations<P, F>(&self, f: F) -> Vec<P> where
F: Fn(&O) -> P,

Returns the set of obligations that are in a pending state.

pub fn process_obligations<P>(
&mut self,
processor: &mut P,
do_completed: DoCompleted
) -> Outcome<O, P::Error> where
P: ObligationProcessor<Obligation = O>,

Performs a pass through the obligation list. This must be called in a loop until outcome.stalled is false.

This cannot be unrolled (presently, at least).