Struct rustc_data_structures::obligation_forest::Outcome

pub struct Outcome<O, E> {
    pub completed: Vec<O>,
    pub errors: Vec<Error<O, E>>,
    pub stalled: bool,
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Fields
completed: Vec<O>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Obligations that were completely evaluated, including all (transitive) subobligations.

errors: Vec<Error<O, E>>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Backtrace of obligations that were found to be in error.

stalled: bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

If true, then we saw no successful obligations, which means there is no point in further iteration. This is based on the assumption that when trait matching returns Err or Ok(None), those results do not affect environmental inference state. (Note that if we invoke process_obligations with no pending obligations, stalled will be true.)