rustc_data_structures::obligation_forest
Trait OutcomeTrait

pub trait OutcomeTrait {
    type Error;
    type Obligation;

    // Required methods
    fn new() -> Self;
    fn record_completed(&mut self, outcome: &Self::Obligation);
    fn record_error(&mut self, error: Self::Error);
}

This trait allows us to have two different Outcome types:

the normal one that does as little as possible

one for tests that does some additional work and checking

Required Associated Types

type Error

type Obligation

Required Methods

fn new() -> Self

fn record_completed(&mut self, outcome: &Self::Obligation)

fn record_error(&mut self, error: Self::Error)

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.