Trait rustc_data_structures::owning_ref::IntoErasedSendSync

pub unsafe trait IntoErasedSendSync<'a>: Send + Sync {
    type Erased: Send + Sync;
    fn into_erased_send_sync(self) -> Self::Erased;
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Helper trait for erasing the concrete type of what an owner derferences to, for example Box<T> -> Box<Erased + Send + Sync>. This would be unneeded with higher kinded types support in the language.

Associated Types

type Erased: Send + Sync

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Owner with the dereference type substituted to Erased + Send + Sync.

Required Methods

fn into_erased_send_sync(self) -> Self::Erased

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Perform the type erasure.