Struct rustc_data_structures::profiling::SelfProfiler

pub struct SelfProfiler {
    profiler: Profiler,
    event_filter_mask: EventFilter,
    string_cache: RwLock<FxHashMap<String, StringId>>,
    query_event_kind: StringId,
    generic_activity_event_kind: StringId,
    incremental_load_result_event_kind: StringId,
    incremental_result_hashing_event_kind: StringId,
    query_blocked_event_kind: StringId,
    query_cache_hit_event_kind: StringId,
    artifact_size_event_kind: StringId,
}

Fields
profiler: Profilerevent_filter_mask: EventFilterstring_cache: RwLock<FxHashMap<String, StringId>>query_event_kind: StringIdgeneric_activity_event_kind: StringIdincremental_load_result_event_kind: StringIdincremental_result_hashing_event_kind: StringIdquery_blocked_event_kind: StringIdquery_cache_hit_event_kind: StringIdartifact_size_event_kind: StringId
Implementations

impl SelfProfiler

pub fn new(
output_directory: &Path,
crate_name: Option<&str>,
event_filters: Option<&[String]>,
counter_name: &str
) -> Result<SelfProfiler, Box<dyn Error + Send + Sync>>

pub fn alloc_string<STR: SerializableString + ?Sized>(&self, s: &STR) -> StringId

Allocates a new string in the profiling data. Does not do any caching or deduplication.

pub fn get_or_alloc_cached_string<A>(&self, s: A) -> StringIdwhere
A: Borrow<str> + Into<String>,

Gets a StringId for the given string. This method makes sure that any strings going through it will only be allocated once in the profiling data.

pub fn map_query_invocation_id_to_string(
&self,
from: QueryInvocationId,
to: StringId
)

pub fn bulk_map_query_invocation_id_to_single_string<I>(
&self,
from: I,
to: StringId
)where
I: Iterator<Item = QueryInvocationId> + ExactSizeIterator,

pub fn query_key_recording_enabled(&self) -> bool

pub fn event_id_builder(&self) -> EventIdBuilder<'_>