rustc_data_structures::sso
Struct SsoHashSet

pub struct SsoHashSet<T> {
    map: SsoHashMap<T, ()>,
}

Small-storage-optimized implementation of a set.

Stores elements in a small array up to a certain length and switches to HashSet when that length is exceeded.

Fields
map: SsoHashMap<T, ()>
Implementations

impl<T> SsoHashSet<T>

pub fn new() -> Self

Creates an empty SsoHashSet.

pub fn with_capacity(cap: usize) -> Self

Creates an empty SsoHashSet with the specified capacity.

pub fn clear(&mut self)

Clears the set, removing all values.

pub fn capacity(&self) -> usize

Returns the number of elements the set can hold without reallocating.

pub fn len(&self) -> usize

Returns the number of elements in the set.

pub fn is_empty(&self) -> bool

Returns true if the set contains no elements.

pub fn iter(&self) -> impl Iterator<Item = &T>

An iterator visiting all elements in arbitrary order. The iterator element type is &'a T.

pub fn drain(&mut self) -> impl Iterator<Item = T> + '_

Clears the set, returning all elements in an iterator.

impl<T: Eq + Hash> SsoHashSet<T>

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the SsoHashSet. The collection may reserve more space to avoid frequent reallocations.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the set as much as possible. It will drop down as much as possible while maintaining the internal rules and possibly leaving some space in accordance with the resize policy.

pub fn retain<F>(&mut self, f: F)
where F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

pub fn take(&mut self, value: &T) -> Option<T>

Removes and returns the value in the set, if any, that is equal to the given one.

pub fn get(&self, value: &T) -> Option<&T>

Returns a reference to the value in the set, if any, that is equal to the given value.

pub fn insert(&mut self, elem: T) -> bool

Adds a value to the set.

Returns whether the value was newly inserted. That is:

If the set did not previously contain this value, true is returned.

If the set already contained this value, false is returned.

pub fn remove(&mut self, value: &T) -> bool

Removes a value from the set. Returns whether the value was present in the set.

pub fn contains(&self, value: &T) -> bool

Returns true if the set contains a value.