Module rustc_data_structures::stable_hasher

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Structs

StableHasher 	 [
Experimental
]
When hashing something that ends up affecting properties like symbol names, we want these symbol names to be calculated independently of other factors like what architecture you're compiling from.

StableVec 	 [
Experimental
]
A vector container that makes sure that its items are hashed in a stable order.

Traits

HashStable 	 [
Experimental
]
Something that implements HashStable<CTX> can be hashed in a way that is stable across multiple compilation sessions.

StableHasherResult 	 [
Experimental
]

StableHashingContextProvider 	 [
Experimental
]
Something that can provide a stable hashing context.

ToStableHashKey 	 [
Experimental
]
Implement this for types that can be turned into stable keys like, for example, for DefId that can be converted to a DefPathHash. This is used for bringing maps into a predictable order before hashing them.

Functions

hash_stable_hashmap 	 [
Experimental
]