Struct rustc_data_structures::stable_hasher::StableHasher

pub struct StableHasher {
    state: SipHasher128,
}

When hashing something that ends up affecting properties like symbol names, we want these symbol names to be calculated independently of other factors like what architecture youโre compiling from.

To that end we always convert integers to little-endian format before hashing and the architecture dependent isize and usize types are extended to 64 bits if needed.

Fields
state: SipHasher128
Implementations

impl StableHasher

pub fn new() -> Self

pub fn finish<W: StableHasherResult>(self) -> W

impl StableHasher

pub fn finalize(self) -> (u64, u64)