Struct rustc_data_structures::stable_hasher::StableHasher

pub struct StableHasher<W> {
    state: SipHasher128,
    bytes_hashed: u64,
    width: PhantomData<W>,
}

When hashing something that ends up affecting properties like symbol names, we want these symbol names to be calculated independently of other factors like what architecture you're compiling from.

To that end we always convert integers to little-endian format before hashing and the architecture dependent isize and usize types are extended to 64 bits if needed.

Fields
state: SipHasher128 bytes_hashed: u64 width: PhantomData<W>
Methods

impl<W: StableHasherResult> StableHasher<W>

pub fn new() -> Self

pub fn finish(self) -> W

impl<W> StableHasher<W>

pub fn finalize(self) -> (u64, u64)

pub fn bytes_hashed(&self) -> u64