Trait rustc_data_structures::stable_hasher::HashStable

pub trait HashStable<CTX> {
    fn hash_stable<W: StableHasherResult>(
        &self, 
        hcx: &mut CTX, 
        hasher: &mut StableHasher<W>
    );
}

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Something that implements HashStable<CTX> can be hashed in a way that is stable across multiple compilation sessions.

Required Methods

fn hash_stable<W: StableHasherResult>(
&self,
hcx: &mut CTX,
hasher: &mut StableHasher<W>
)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, and unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?