Trait rustc_data_structures::stable_hasher::ToStableHashKey

pub trait ToStableHashKey<HCX> {
    type KeyType: Ord + Sized + HashStable<HCX>;

    fn to_stable_hash_key(&self, hcx: &HCX) -> Self::KeyType;
}

Implement this for types that can be turned into stable keys like, for example, for DefId that can be converted to a DefPathHash. This is used for bringing maps into a predictable order before hashing them.

Required Associated Types

type KeyType: Ord + Sized + HashStable<HCX>

Required Methods

fn to_stable_hash_key(&self, hcx: &HCX) -> Self::KeyType