Module rustc_data_structures::sync

This module defines types which are thread safe if cfg!(parallel_queries) is true.

Lrc is an alias of either Rc or Arc.

Lock is a mutex. It internally uses parking_lot::Mutex if cfg!(parallel_queries) is true, RefCell otherwise.

RwLock is a read-write lock. It internally uses parking_lot::RwLock if cfg!(parallel_queries) is true, RefCell otherwise.

LockCell is a thread safe version of Cell, with set and get operations. It can never deadlock. It uses Cell when cfg!(parallel_queries) is false, otherwise it is a Lock.

MTLock is a mutex which disappears if cfg!(parallel_queries) is false.

rustc_erase_owner! erases a OwningRef owner into Erased or Erased + Send + Sync depending on the value of cfg!(parallel_queries).

Re-exports

pub use self::serial_join as join;

pub use self::serial_scope as scope;

Structs

Lock

LockCell

LockGuard
A wrapper type for a mutably borrowed value from a RefCell<T>.

Lrc
A single-threaded reference-counting pointer. 'Rc' stands for 'Reference Counted'.

MTLock

Once
A type whose inner value can be written once and then will stay read-only

OneThread
A type which only allows its inner value to be used in one thread. It will panic if it is used on multiple threads.

ReadGuard
Wraps a borrowed reference to a value in a RefCell box. A wrapper type for an immutably borrowed value from a RefCell<T>.

RwLock

SerialScope

Weak
Weak is a version of Rc that holds a non-owning reference to the managed value. The value is accessed by calling upgrade on the Weak pointer, which returns an Option<Rc<T>>.

WriteGuard
A wrapper type for a mutably borrowed value from a RefCell<T>.

Traits

HashMapExt

ParallelIterator
An interface for dealing with iterators.

Send

Sync

Functions

assert_send_sync_val

assert_send_val

assert_sync

par_iter

serial_join

serial_scope

Type Definitions

MetadataRef