Struct rustc_data_structures::sync::Once

pub struct Once<T>(_, _);

A type whose inner value can be written once and then will stay read-only

Methods

impl<T> Once<T>

pub fn new() -> Self

Creates an Once value which is uninitialized

pub fn into_inner(self) -> Option<T>

Consumes the value and returns Some(T) if it was initialized

pub fn try_set(&self, value: T) -> Option<T>

Tries to initialize the inner value to value. Returns None if the inner value was uninitialized and value was consumed setting it otherwise if the inner value was already set it returns value back to the caller

pub fn try_set_same(&self, value: T) -> Option<T> where
T: Eq,

Tries to initialize the inner value to value. Returns None if the inner value was uninitialized and value was consumed setting it otherwise if the inner value was already set it asserts that value is equal to the inner value and then returns value back to the caller

pub fn set(&self, value: T)

Tries to initialize the inner value to value and panics if it was already initialized

pub fn init_locking<F: FnOnce() -> T>(&self, f: F) -> bool

Tries to initialize the inner value by calling the closure while ensuring that no-one else can access the value in the mean time by holding a lock for the duration of the closure. If the value was already initialized the closure is not called and false is returned, otherwise if the value from the closure initializes the inner value, true is returned

pub fn init_nonlocking<F: FnOnce() -> T>(&self, f: F) -> Option<T>

Tries to initialize the inner value by calling the closure without ensuring that no-one else can access it. This mean when this is called from multiple threads, multiple closures may concurrently be computing a value which the inner value should take. Only one of these closures are used to actually initialize the value. If some other closure already set the value, we return the value our closure computed wrapped in a Option. If our closure set the value, None is returned. If the value is already initialized, the closure is not called and None is returned.

pub fn init_nonlocking_same<F: FnOnce() -> T>(&self, f: F) -> Option<T> where
T: Eq,

Tries to initialize the inner value by calling the closure without ensuring that no-one else can access it. This mean when this is called from multiple threads, multiple closures may concurrently be computing a value which the inner value should take. Only one of these closures are used to actually initialize the value. If some other closure already set the value, we assert that it our closure computed a value equal to the value already set and then we return the value our closure computed wrapped in a Option. If our closure set the value, None is returned. If the value is already initialized, the closure is not called and None is returned.

pub fn try_get(&self) -> Option<&T>

Tries to get a reference to the inner value, returns None if it is not yet initialized

Important traits for &'_ mut I
impl<'_, I> Iterator for &'_ mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn get(&self) -> &T

Gets reference to the inner value, panics if it is not yet initialized

Important traits for &'_ mut I
impl<'_, I> Iterator for &'_ mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn borrow(&self) -> &T

Gets reference to the inner value, panics if it is not yet initialized