rustc_data_structures::sync
Module worker_local

Structs

Registry

Represents a list of threads which can access worker locals.

RegistryData 🔒

RegistryId 🔒

A pointer to the RegistryData which uniquely identifies a registry. This identifier can be reused if the registry gets freed.

ThreadData 🔒

WorkerLocal

Holds worker local values for each possible thread in a registry. You can only access the worker local value through the Deref impl on the registry associated with the thread it was created on. It will panic otherwise.

Constants

REGISTRY 🔒

The registry associated with the thread. This allows the WorkerLocal type to clone the registry in its constructor.

THREAD_DATA 🔒

A thread local which contains the identifier of REGISTRY but allows for faster access. It also holds the index of the current thread.