Struct rustc_data_structures::transitive_relation::TransitiveRelation

pub struct TransitiveRelation<T: Clone + Debug + Eq + Hash + Clone> { /* fields omitted */ }

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Methods

impl<T: Clone + Debug + Eq + Hash + Clone> TransitiveRelation<T>

pub fn new() -> TransitiveRelation<T>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn is_empty(&self) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

pub fn maybe_map<F, U>(&self, f: F) -> Option<TransitiveRelation<U>> where
F: FnMut(&T) -> Option<U>,
U: Clone + Debug + Eq + Hash + Clone,

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Applies the (partial) function to each edge and returns a new relation. If f returns None for any end-point, returns None.

pub fn add(&mut self, a: T, b: T)

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Indicate that a < b (where < is this relation)

pub fn contains(&self, a: &T, b: &T) -> bool

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Check whether a < target (transitively)

pub fn reachable_from(&self, a: &T) -> Vec<&T>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Thinking of x R y as an edge x -> y in a graph, this returns all things reachable from a.

Really this probably ought to be impl Iterator<Item=&T>, but I'm too lazy to make that work, and -- given the caching strategy -- it'd be a touch tricky anyhow.

pub fn postdom_upper_bound(&self, a: &T, b: &T) -> Option<&T>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Picks what I am referring to as the "postdominating" upper-bound for a and b. This is usually the least upper bound, but in cases where there is no single least upper bound, it is the "mutual immediate postdominator", if you imagine a graph where a < b means a -> b.

This function is needed because region inference currently requires that we produce a single "UB", and there is no best choice for the LUB. Rather than pick arbitrarily, I pick a less good, but predictable choice. This should help ensure that region inference yields predictable results (though it itself is not fully sufficient).

Examples are probably clearer than any prose I could write (there are corresponding tests below, btw). In each case, the query is postdom_upper_bound(a, b):

// returns Some(x), which is also LUB
a -> a1 -> x
           ^
           |
b -> b1 ---+

// returns Some(x), which is not LUB (there is none)
// diagonal edges run left-to-right
a -> a1 -> x
  \/       ^
  /\       |
b -> b1 ---+

// returns None
a -> a1
b -> b1

pub fn mutual_immediate_postdominator<'a>(
&'a self,
mubs: Vec<&'a T>
) -> Option<&'a T>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Viewing the relation as a graph, computes the "mutual immediate postdominator" of a set of points (if one exists). See postdom_upper_bound for details.

pub fn minimal_upper_bounds(&self, a: &T, b: &T) -> Vec<&T>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Returns the set of bounds X such that:

a < X and b < X

there is no Y != X such that a < Y and Y < X

except for the case where X < a (i.e., a strongly connected component in the graph). In that case, the smallest representative of the SCC is returned (as determined by the internal indices).

Note that this set can, in principle, have any size.

pub fn parents(&self, a: &T) -> Vec<&T>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

Given an element A, returns the maximal set {B} of elements B such that

A != B

A R B is true

for each i, j: B[i] R B[j] does not hold

The intuition is that this moves "one step up" through a lattice (where the relation is encoding the <= relation for the lattice). So e.g. if the relation is -> and we have

a -> b -> d -> f
|              ^
+--> c -> e ---+

then parents(a) returns [b, c]. The postdom_parent function would further reduce this to just f.

pub fn postdom_parent(&self, a: &T) -> Option<&T>

🔬 This is a nightly-only experimental API. (rustc_private)

this crate is being loaded from the sysroot, an unstable location; did you mean to load this crate from crates.io via Cargo.toml instead?

A "best" parent in some sense. See parents and postdom_upper_bound for more details.