Trait rustc_data_structures::unify::UnifyKey

pub trait UnifyKey: Copy + Clone + Debug + PartialEq {
    type Value: Clone + PartialEq + Debug;
    fn index(&self) -> u32;
    fn from_index(u: u32) -> Self;
    fn tag(k: Option<Self>) -> &'static str;
}

Unstable (rustc_private)

This trait is implemented by any type that can serve as a type variable. We call such variables unification keys. For example, this trait is implemented by IntVid, which represents integral variables.

Each key type has an associated value type V. For example, for IntVid, this is Option<IntVarValue>, representing some (possibly not yet known) sort of integer.

Clients are expected to provide implementations of this trait; you can see some examples in the test module.

Associated Types

type Value: Clone + PartialEq + Debug

Unstable (rustc_private)

Required Methods

fn index(&self) -> u32

Unstable (rustc_private)

fn from_index(u: u32) -> Self

Unstable (rustc_private)

fn tag(k: Option<Self>) -> &'static str

Unstable (rustc_private)