Trait rustc_data_structures::unord::UnordCollection

trait UnordCollection { }

A marker trait specifying that Self can consume UnordItems<_> without exposing any internal ordering.

Note: right now this is just a marker trait. It could be extended to contain some useful, common methods though, like len, clear, or the various kinds of to_sorted.